Fix transaction modal "details" button
Previously, when clicking on a transaction to open the transaction modal (TxDetailsModal.js) and then clicking on the "details" button the explorer wouldn't open correctly as there were two "/" after the domain name rather than one. The error was in the CoinData.js file where some of the explorer URLs had an unwanted "/" at the end.